+++ CHANGELOG.md
@@ -35,6 +35,7 @@ RELEASING:
 ## [unreleased]
 
 ### Added
+- TopoJSON graph export ([#1926](https://github.com/GIScience/openrouteservice/pull/1926))
 
 ### Changed